pyproject.toml
[tool.poetry]
name = "scoutnet2airkey"
version = "0.0.0"
description = "Scoutnet to EVVA Airkey integration"
authors = ["Jakob Schlyter <jakob@schlyter.se>"]

[tool.poetry.dependencies]
python = "^3.9"
certifi = "^2020.12.5"
six = "^1.15.0"
python-dateutil = "^2.8.1"
urllib3 = "^1.26.4"
requests = "^2.25.1"
PyYAML = "^5.4.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
